+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.67.2
 - Read X-INSTANA-... headers case-insensitive in amqp instrumentation.
 - Fix handling of Fastify routes with an beforeHandler array.